Defer I/O until the data is needed
I consider it bad manners to perform I/O (reads and write) of unknown latency in the `__init__` method. If you are initializing a big herd of these objects, you have to wait for all of them to load their `entries` attribute until you can use them. A side-effect of this property pattern is in-memory caching of the results.